add: Book 모델에 id 추가

+++ feature/bookdetail/src/main/java/org/sopt/bookdetail/Route.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 sealed interface Route {
     @Serializable
-    data class BookDetail(val bookId: Int) : Route
+    data class BookDetail(val bookId: Long) : Route
 }
